(GH-2511) Only spin while executing run_* plan functions
Previously the spinner would spin for the entirety of a plan or task
run, starting and stopping from the Bolt CLI manager. This meant that if
the user was prompted during a plan run the spinner would spin in the
area of the terminal the user was supposed to input text. While this
didn't write to stdin or impact user input, it was annoying and
definitely wrong.

This modifies the spinner to be started and stopped from the executor,
impacting calls from the command line and plan functions. The executor
only starts and stops the spinner as part of `with_node_logging()`,
which affects `run_*` functions and `wait_until_available`. The spinner
still spins for module management and project updating commands, which
are unaffected by this change. The spinner will only be started and
stopped if Bolt is not executing inside a `parallelize` block - instead
the parallelize block itself (specifically `Bolt::Executor#round_robin`)
will manage starting and stopping the spinner. This is so that the first
element in the array that finishes doesn't signal to stop spinning
before other functions have completed, but the spinner will spin for the
entire parallelize block as expected.

Test cases:
- [ ] The spinner spins while executing executor plan functions
- [ ] The spinner spins while executing `bolt command run`, `bolt script
  run`, `bolt task run`, `bolt file upload`, and `bolt file download`,
  and module management commands
- [ ] The spinner spins normally while running multiple functions in a
  parallelize block across multiple targets
- [ ] Spinner does not spin when running other plan functions or Bolt
  commands (in particular when prompting for user input)

!bug

* **Only spin while executing run_* plan functions** ([#2511](https://github.com/puppetlabs/bolt/issues/2511))
  Bolt will now only print the spinner while executing `run_*`,
  `file_upload`, `file_download`, and`wait_until_available` plan
  functions. It also now spins while running those functions equivalent
  commandline commands. This prevents the spinner from spinning while
  prompting for output from a plan.