(GH-2522) Add 'resolve' key to modules config
This adds support for a new `resolve` key for both Forge and git module
specs in `bolt-project.yaml`. It also adds support for a `name` key for
git module specs, which is required when setting the `resolve` key to
`false`.

When `resolve` is `false`, Bolt will skip dependency resolution for the
module. This allows users to include modules that would otherwise break
Bolt's dependency resolution in their Bolt projects, avoiding the need
to manually manage a Puppetfile. This includes modules with broken
metadata and git modules hosted in a repository other than a public
GitHub repository.

If a unresolved module shares the name of another module's dependency,
Bolt will error indicating that there is a name conflict. This is in
line with r10k's behavior when installing a Puppetfile that has
duplicate module names.

!feature

* **Add `resolve` key for Forge and git module specifications**
  ([#2522](https://github.com/puppetlabs/bolt/issues/2522))

  Forge and git module specifications in `bolt-project.yaml` now support
  a `resolve` key. When setting `resolve: false`, Bolt will skip
  dependency resolution for the module, allowing users to include
  modules with broken metadata or modules hosted in a repository other
  than a public GitHub repository in their project configuration.