(GH-2543) Block publishing events while prompting and parallel
This will block the executor from continuing to publish events until the
user has entered input for a prompt. This prevents prompts inside of
parallelize blocks from being interleaved with output from the block
continuing to execute.

This also adds a known issue that output from `resolve_reference` that
uses `prompt` inside a parallel block may not be a good user experience.

Lastly, this adjusts the spinner while running in parallel to
pessimistically spin - that is, only spin if it's started by a
long-running process and not stopped. This includes sending the `spin:
stop` message to the spinner at any point that a
non-long-running-process (like prompt) might begin.

Closes #2543

!bug

* **Don't continue executing parallel block when prompting** ([#2543](https://github.com/puppetlabs/bolt/issues/2543))
  Bolt will now pause printing messages from parallel blocks when
  prompting the user for input, to avoid confusing printing to the screen.