(GH-2542) Add disable-warnings config option to disable warnings
This adds a new `disable-warnings` config option that is available in
both `bolt-project.yaml` and `bolt-defaults.yaml`. This option accepts
an array of warning IDs that are used to suppress warnings issued by
Bolt. If a warning would otherwise be issued, and its ID is under the
`disable-warnings` option, Bolt will not print the warning to either the
CLI or a log file.

If `disable-warnings` is configured in multiple config layers, Bolt will
merge the arrays between config layers.

To enable this feature, a new `Bolt::Logger.warn` method was added that
accepts an ID and message. If the ID is present in the
`disable-warnings` config, the message will not be logged.

Lastly, this adds a queuing mechanic to `Bolt::Logger`. This work
includes the new methods `deprecate`, `deprecate_once`, `info`, and
`debug`; and updates the existing `warn_once` method. These methods now
log messages via the `log` method. If `Bolt::Logger` has been
configured, the messages are immediately logged. Otherwise, the messages
are added to a queue to be logged once `Bolt::Logger` is configured. The
`flush_queue` method can be called to flush the message queue and log
all messages in the queue.

!feature

* **Suppress warnings with `disable-warnings` config option**
  ([#2542](https://github.com/puppetlabs/bolt/issues/2542))

  The `disable-warnings` configuration option accepts an array of
  warning IDs that are used to suppress warnings in both the CLI and log
  files. This configuration option is supported in both
  `bolt-project.yaml` and `bolt-defaults.yaml`.