(GH-102) Prototype streaming remote output
This adds a new configuration option `stream` that will enable
streaming output back to the console while Bolt is executing actions on
remote targets.

Streaming works by registering a new Ruby logger named 'stream' that is
not additive (so does not send log messages to the parent logger, which
prevents messages from being logged to files or to other Bolt logging
streams). It configures the stream logger to only print to the console
without much formatting. The stream logger is only created if `stream`
is configured. Bolt then prints output to the stream logger as it's
read, in addition to sending messages to the Result object to be
returned and printed when the action is finished.

Streaming messages are prepended with the target's safe name and the
stream (stdout or stderr) the message came from. Streaming messages are
generally printed twice, once when streaming and again as part of the
result message.

This feature is experimental.

!feature

* **Add 'stream' option** ([#102](https://github.com/puppetlabs/bolt/issues/102))

  **This option is experimental**. New `stream` configuration option and
  `--stream` CLI flag enable streaming output from Bolt actions as they are running.