(GH-2548) Update tests to run in Ruby 2.7
This updates our test infrastructure to run in Ruby 2.7 in order to
match the version Bolt ships with and runs CI in. This additionally
updates the minimum Ruby version of Bolt to be 2.7, so that Bolt doesn't
break in unexpected ways due to using a lower Ruby version than
expected.

!feature

* **Bolt now ships with Ruby 2.7**

  Bolt will now ship with, and run in, Ruby 2.7, upgraded from 2.5.