(docs) Include namespace in Brew install instructions
The puppet-bolt brew package is hosted in a [Puppet homebrew
repo](https://github.com/puppetlabs/homebrew-puppet) and is not shipped
as part of the core Homebrew repo ([see this for
context](https://github.com/Homebrew/homebrew-core/pull/28439)). Since
it's in our own namespace and not the core namespace, we need to include
that namespace in the installation instructions.

!no-release-note