(docs) Clean up install docs
This cleans up the 'Install Bolt' docs page. It adds a table to quickly
reference supported platforms and also moves install commands for Linux
distributions under separate headers for easy linking. It also breaks
out the analytics section into its own dedicated page so it's easier to
find the information. Lastly, the navigation was updated to have
"Installing Bolt" as a navigation title.

!no-release-note