(GH-2311) Cleanup LXD transport work
This cleans up the awesome work on the LXD transport to be more in line
with the Bolt codebase, and simplify the transport to just the needed
functions.

For now the LXD transport will only accept `cleanup` and `tmpdir`
configuration options. `remote` will be added later, in order to limit
scope for this initial work.

The LXD transport is experimental, and will raise a warning stating so
each time Bolt is run. Like all warnings this warning can be silenced
with configuration.

!no-release-note