(PE-31142) Serialize sensitive output from task Result
This commit updates the `to_data` method for the `Result` class to serialize sensitive task output by calling the `to_s` value explictly on the sensitive output. The purpose of the `to_data` method is to represent a `Result` as the puppet `Data` type. A `Sensitive` type is not puppet `Data` and therefore the method was insufficient for data serialization leading to (for example JSON) serialization bugs when trying to get a simple representation of a `Result`.

!bug

* **Serialize Sensitive task output for result to_data method**

  Previously the `to_data` method on a `Result` object would not
  transform `Sensitive` task output. Now the `to_data` method
  serializes the output by calling the `to_s` method on `Sensitive`
  output (which will simply print the "value redacted" message).