(GH-2652) Support Puppet paths in 'bolt script run'
This adds support for using Puppet paths (`<MODULE NAME>/<FILE NAME>`)
in `bolt script run` and `Invoke-BoltScript`. The default behavior is to
still locate the script file using an absolute/relative path, but Bolt
will now fall back to checking if the path is a Puppet path if that
fails.

!feature

* **Support Puppet paths when running scripts from the CLI**
  ([#2652](https://github.com/puppetlabs/bolt/issues/2652))

  You can now use Puppet paths (`<MODULE NAME>/<FILE NAME>`) to specify
  the path to a script when running `bolt script run` or
  `Invoke-BoltScript`.