(GH-2629) Add top-level metaparameter keys to YAML plan steps
This adds keys to most YAML plan steps that correspond to plan function
metaparameters. For example, the script step now accepts an `env_vars`
key that corresponds to the `_env_vars` metaparameter, and allows users
to set environment variables on a target when executing the script.

This also fixes a bug where targets listed under the `targets` key in a
plan step were not being passed to the plan. Targets specified in this
manner are now passed to the plan correctly.

Lastly, this includes a refactoring of the `Bolt::PAL::YamlPlan::Step`
classes and the `Bolt::PAL::YamlPlan::Evaluator` class. The step classes
now include methods for formatting options and arguments for plan
function calls as well as an `#evaluate` method that handles evaluating
the step body and calling the step's plan function. The evaluator now
calls out to each step's `#evaluate` function to evaluate the step
instead of including separate logic for evaluating each step and
formatting the plan function call. The motivation for this refactoring
was to keep all of the logic for a given step in the step class, as
opposed to spreading it across multiple classes with duplicate logic.

!feature

* **Support metaparameters as top-level keys in YAML plan steps**
  ([#2629](https://github.com/puppetlabs/bolt/issues/2629))

  YAML plan steps now support metaparameters as top-level keys. For
  example, the `script` step supports an `env_vars` key which accepts a
  hash of environment variables to set on the target when running the
  script.

!bug

* **Run YAML plan `plan` steps with `targets` key**

  YAML plans that have a `plan` step with a top-level `targets` key now
  pass the targets to the plan.