(maint) Prefer 'might' to 'may' where appropriate
We learned recently that 'might' is preferable to 'may' when expressing
possibility or making suggestions. This replaces any use of 'may' that
can be replaced with 'might' in messages printed by Bolt and the
documentation. It ignores cases where 'may' refers to ability (ie. "You
may include lowercase letters"), and comments in the Bolt codebase.

!no-release-note