(maint) Don't include password for Windows target
The Windows testing password includes characters that aren't allowed in
target uris that are looked up in the inventory file. When we use the
full target uri (including password) from the `BoltSpec::Conn` module
along with a project created by `BoltSpec::Project#with_project` which
laid down an empty inventoryfile, Bolt would try to lookup the uri in
the empty inventory and fail.

This makes two changes to address this:
- First, it modifies the failing test to use `--password` instead of
  including the password in the URI.
- Second, it modifies `with_project` to only create an inventoryfile if
  any inventory data is passed to the function. This will prevent Bolt
  from trying to lookup targets in the inventory unnecessarily.

!no-release-note