(GH-2653) Merge stdout and stderr for commands and scripts
This updates the `bolt command|script run` commands to display
merged output from stdout and stderr on the CLI. When Bolt logs the
output from these two streams, it also adds each line to a
`merged_output` value in the order the lines are received. This
`merged_output` value is passed to the result for the command or script,
and can be accessed from both plans and the JSON output format. Lastly,
the human outputter has been updated to display the merged output by
default, resulting in output resembling what a user would see if they
ran the command or script directly on the target.

!feature

* **Display merged stdout and stderr output for commands and scripts**
  ([#2653](https://github.com/puppetlabs/bolt/issues/2653))

  The `bolt command|script run` commands and `Invoke-BoltCommand|Script`
  cmdlets now display merged output from stdout and stderr in the CLI.
  This merged output is also available to the `Result` object in plans
  and in the JSON output format under the `merged_output` key.