(POC) run_container plan function
This POC introduces the `run_container()` plan function. The function
accepts the name of an image (optionally with  a registry prepended) to
run, and a hash of options to apply when running the container. For this
POC, Docker is the only supported container engine. The function returns
a new `ContainerResult` object that contains a `value`, `status`, and
`object`. This function is supported on both *nix and Windows systems.

Currently, supported options are:
* `cmd` - A command to run in the container
* `env_vars` - A hash of environment variables to set in the container
* `ports` - A hash of ports to publish from the container to the host
* `rm` - A boolean, whether to remove the container when it exits
* `volumes` - A hash of directories on the host and their corresponding
  directories in the container to mount into the container

If you have requests for more options to add, let us know in the
comments, in a new issue, or in Slack (and be sure to smash that 'star'
button 😉).

!feature

* **run_container() plan function** ([#2614](https://github.com/puppetlabs/bolt/issues/2614))

  This introduces a new `run_container()` Puppet plan function that will
  run a container and return its output.