(GH-2680) Do not send task parameters over stdin when using a tty
This fixes a bug with running tasks using the `stdin` input method when
using a `tty`. When the `tty` config option is set, Bolt creates a
wrapper script that executes the task and passes the task's parameters
over stdin. However, Bolt would still invoke this wrapper script on the
target and attempt to pass the parameters over stdin to the wrapper
script, even though the wrapper script already included the task
parameters. This resulted in the parameters being passed to the wrapper
script to print to stdout and be incorrectly returned as part of the
task's output.

!bug

* **Do not send task parameters over stdin when using a tty**
  ([#2680](https://github.com/puppetlabs/bolt/issues/2680))

  Tasks with a `stdin` input method that are run on targets with `tty:
  true` configuration no longer return the task's parameters as part of
  the task output. Previously, Bolt was sending these parameters to the
  task twice, causing them to be printed to standard out (stdout) and
  returned in the task output.