(docs) Correct modulepath precedence documentation
The Modules overview document listed the wrong precedence of paths on
the modulepath, stating that `<project>/.modules/` would override
`<project>/modules`. This corrects the order to state that any modules
in `<project>/modules` will be used if they have the same name as
`<project>/.modules` - basically that user-created modules will override
installed modules.

!no-release-note