(GH-2719) Correctly show task info if task parameters don't have a type
Previously, Bolt assumed that task parameters included the `type` key
when checking to see if the `type` was optional to format it for
printing in `bolt task show` output. If the task included parameters
that didn't specify a type, Bolt would stacktrace. This now only adds
the `optional` parameter formatting if the parameter has a type defined
and that type includes 'Optional', otherwise it formats the type as
required. This prevents Bolt from stacktracing.

!bug

* **Don't stacktrace when showing tasks that include untyped parameters** ([#2719](https://github.com/puppetlabs/bolt/issues/2719))
  Bolt will now correctly show task details for tasks that include
  parameters that do not specify a type, instead of stacktracing.