(bug) Handle malformed `_error` values in task result in Orchestrator
Previously, we assumed that if a task returned with an `_error` key that
it would be a hash and include the `details` key, and that that
`details` key would also have a hash value. Any of these assumptions
might be wrong. This amends the Orchestrator transport to handle
`_error` values that:
* Are not a hash, by placing the provided value under the `message` key
* Is a hash that doesn't have the `details` key by setting the details
  key to an empty hash
* Is a hash with a `details` key that is not itself a hash by not trying
  to merge our own keys into it.

!bug

* **Handle malformed `_error` values in task results in Orchestrator**

  Bolt now handles `_error` from task results in the Orchestrator
  transport when the value of the key is not a hash, does not include the
  `details` key, or the `details` key is not a hash. Previously Bolt would
  error if any of these conditions was true.