(GH-2692) Support 'run_task_with' in BoltSpec
This adds support for testing plans that use the `run_task_with()` plan
function to BoltSpec. When BoltSpec runs a plan with this function, it
will execute a separate task run for each target. Users making
assertions about tasks run with `run_task_with()` should make an
assertion about each individual target that the task runs against, or an
assertion about the number of times that the task is run. This includes
an example in the 'Testing plans' documentation that details how to test
these types of plans.

!feature

* **Test plans that use `run_task_with()` plan function in BoltSpec*
  ([#2692](https://github.com/puppetlabs/bolt/issues/2692))

  Plans that use the `run_task_with()` plan function can now be tested
  with BoltSpec.