(PE-31344) Validate puppet_library plugin hook for connect
The puppet_library plugin hook can be configured to run an arbitrary task (following the task plugin spec) for connect. Currently only the task plugin (ie not module plugin) is supported. For the validation plan, bolt uses the default puppet_agent module based plugin where connect will specify this same default using the task based plugin notation. The plan validates that either the default is being used or that a task based plugin is being used. The other current limitation is that for connect every target must have the same plugin configured. The validation plan validates this.

Finally, the test file for this module has been renamed so the tests contained in it will actually be run in the rake task triggered by github actions.

!no-release-note