(GH-2550) Publish PuppetBolt module to PowerShell Gallery
This updates the developer documentation for releasing Bolt to include
steps for manually releasing a new version of the PuppetBolt PowerShell
module. It also updates the `pwsh:generate_module` rake task to generate
files for the module in a `PuppetBolt` directory, which contains only
the files that are shipped with the module.

!no-release-note