(GH-2737) Deprecate dotted fact names
This updates Bolt to log a deprecation warning when it detects a target
that has facts with dotted names. This warning is logged whether a
target is loaded with a dotted fact name (either in the inventory or
with `Target.new`) or when adding facts to a target (with
`$target.add_facts`).

!deprecation

* **Deprecate dotted fact names**
  ([#2737](https://github.com/puppetlabs/bolt/issues/2737))

  Dotted fact names (e.g. `foo.bar`) are now deprecated. Bolt issues a
  deprecation warning if it detects a target is loaded with these facts
  or has them added during a plan run.