(GH-2730) Add future.file_paths config option
This introduces a new project-level configuration option `future` which
contains a hash that can enable various features that may introduce
breaking changes in Bolt. The first suboption of `future` is
`file_paths`, which will enable loading scripts from the `scripts/`
subdirectory of a module once #2731 is complete. This doesn't include a
release note, preferring a larger release note once the whole feature is
implemented.

!no-release-note