(hack) Improve 'inventory show' output
This updates the output for `bolt inventory show` and
`Get-BoltInventory` to better match the other `show` commands. Targets,
the inventory source, and the target count are now listed under
colorized headings.

This also updates the `--detail` view so that targets are listed in
alphabetical order and use each target's name as a heading for easy
reading. The target's configuration and data is also now displayed in
YAML in the human outputter, to better match the format of the inventory
file. The `--detail` view also now includes the same inventory source
and target count that are in the undetailed view.

!feature

* **Improved output for `inventory show`**

  The `bolt inventory show` and `Get-BoltInventory` command now display
  `human` output in the same format as other `show` commands.