(hack) Improve plan and task info output
This updates the output for `bolt plan|task show <name>` to better align
with the formatting in other `show` commands. Information is now listed
under colorized headings and parameter information is colorized and
spaced out to improve readability. This also wraps the usage string and
descriptions so each line is a maximum of 80 characters long, which
makes it easier to read long descriptions.

!feature

* **Improved plan and task information output**

  The `bolt plan|task show <name>` and `Get-Bolt(Plan|Task) -Name
  <name>` commands now display `human` output in a similar format to
  other `show` commands.