(maint) Address puppetlabs_spec_helper deprecation warning
The `puppetlabs_spec_helper` ruby library we use as part of testing
embedded Bolt modules was raising a deprecation warning that the default
mocking framework was `mocha` but that to silence the warning the tests
themselves should specify which framework was being used. While I'd
honestly love to switch to `rspec` it's a lot of work to do so for
relatively little gain. For now setting the configuration to `mocha` to
silence the warning is easiest.

!no-release-note