(GH-2756) Do not stack trace when missing project config file
This fixes a bug that arises from running the `bolt module add|install`
command when a project does not have a `bolt-project.yaml` file.
Previously, Bolt would call the undefined method
`Bolt::Project.config_file`, result in a stacktrace. Since this method
was removed in Bolt 3.0, the condition that called this method has been
removed.

!bug

* **Do not stack trace when missing project configuration file**
  ([#2756](https://github.com/puppetlabs/bolt/issues/2756))

  Bolt no longer stack traces when installing modules if the project
  does not have a `bolt-project.yaml` configuration file.