(GH-2731) Load files from specified Puppet paths
This introduces a new syntax for loading files from Puppet modules.
Previously, files in Puppet modules could be loaded if the file existed
in the `files/` subdirectory of a module on the modulepath, and the user
provided `mymodule/file.sh` to load a file located at
`mymodule/files/file.sh`.

The new loading behavior is gated behind a new project-level
configuration option. The option is specified under the `future` key
which is used enable different features that may introduce breaking
changes in Bolt. Right now the only suboption that `future` has is
`file_paths`, which enables the following loading behavior.

The new loading behavior loads with the following precedence depending
on what syntax is provided. Note that files can now be loaded from the
`scripts/` directory of a module in addition to the `files/` directory:

Provided: **module/myscript.sh**
- Load from `files/myscript.sh`
Provided: **module/scripts/myscript.sh**
- Load from `files/scripts/myscript.sh`
- Fall back to `scripts/myscript.sh`
Provided: **module/files/myscript.sh**
- Load from `files/files/myscript.sh`
- Fall back to `files/myscript.sh`

!feature

* **Load files from specified Puppet paths** ([#2731](https://github.com/puppetlabs/bolt/issues/2731))

  If the project-level `future.file_paths` configuration is enabled,
  Puppet files can be loaded using the new loading syntax. For more
  information see https://pup.pt/bolt-loading-files.