(GH-2759) Add 'analytics' config option to disable analytics
This adds a new configuration option called `analytics` that disables
data collection in Bolt when set to `false`. The option is supported in
both `bolt-defaults.yaml` and `bolt-project.yaml`. If this option is set
to `false`, it cannot be overriden in another configuration file. For
example, setting `analytics: true` in `bolt-project.yaml` will not
enable data collection if `analytics: false` is set in
`bolt-defaults.yaml`.

!feature

* **Disable analytics in system, user, and project config files**
  ([#2759](https://github.com/puppetlabs/bolt/issues/2759))

  The new `analytics` configuration option can be used to disable data
  collection in Bolt and is supported in both `bolt-defaults.yaml` and
  `bolt-project.yaml`. Disabling data collection cannot be overridden by
  enabling it in another configuration file.