(GH-2747) Default to showing all targets with 'inventory show'
This updates the `bolt inventory show` and `Get-BoltInventory` commands
to display the 'all' group if a targetting option is not provided to the
command. Additionally, the human outputter now displays additional
information to the user, indicating that they can view specific targets
using a targetting option or view target config and data with
`--detail`, assuming the options are not provided by the user.

!feature

* **Default to showing all targets with `bolt inventory show`**
  ([#2747](https://github.com/puppetlabs/bolt/issues/2747))

  The `bolt inventory show` and `Get-BoltInventory` commands now default
  ot showing all targets in the inventory if a targetting option
  (`--targets`, `--query`, `--rerun`) are not provided.