(bugfix) Ensure all messages are streamed and saved to merged_output
Previously, in the Bash shell class when a command had finished we
collected any leftover messages in a read-loop and saved them to the
appropriate stream to be printed. When we added streaming and
merged_output features we did this for the main read-loop, but not for
the cleanup loop that executes once the thread has finished. This
resulted in some scripts and commands not having their results printed
since we now print merged_output, which was missing messages from the
final stream reading.

Now, any messages read once the thread has finished will be read and
logged to the stream output and saved to `merged_output` so that it can
be printed.

!bug

* **Ensure all messages printed, even after thread finishes**

  Bolt now ensures that all messages from a command or script are
  printed back to the user. Previously, some final messages would be lost
  if they were read after the thread finished executing.