(bugfix) Ensure all messages are streamed and saved to merged_output
In the Bash shell class we collect messages from stdout and stderr in a
few places: when using `run-as` after checking for a prompt, during the
read-loop as the command is executing, and then reading anything that's
unread after the thread exits. When we added streaming and merged_output
features we did this for the main read-loop, but not for the initial
sudo-prompt check or the cleanup loop that executes once the thread has
finished. This resulted in some scripts and commands not having their
results printed since we now print merged_output, which was missing
messages from the final stream reading.

Now, any messages read during the initial sudo-prompt check or once the
thread has finished will be read and logged to the stream output and
saved to `merged_output` so that it can be printed.

!bug

* **Ensure all messages printed, even after thread finishes**

  Bolt now ensures that all messages from a command or script are
  printed back to the user. Previously, some messages would be lost
  if they were read after the thread finished executing or when Bolt had
  been prompted for a sudo password.