(Docs) Optimize images for refactor
We're going through all of the docs and optimizing the images. This
is previously something that was done automatically as part of the web
build, but it adds unnecessary time to the build, and most of the
web post-processing is going away, so we're trying to make sure this is
done in the source instead.

!no-release-note