(GH-2777) Do not error when using metaparameters in YAML plans
This fixes a bug that was causing YAML plan plan and task steps that
specified a metaparameter under the `parameters` key to error. The
correct behavior for Bolt is to only error when the metaparameter is
present as both a top-level key _and_ under the `parameters` key. For
example, specifying the `catch_errors` key and the
`parameters._catch_errors` key.

!bug

* **Do not error when using metaparameters in YAML plans**
  ([#2777](https://github.com/puppetlabs/bolt/issues/2777))

  Bolt no longer errors for YAML plans that include a plan or task step
  that includes an additional option (e.g. `_catch_errors`) under the
  `parameters` key.