(GH-2767) Allow Bolt to run with Ruby 3.0
Previously, our gemspec limited the Ruby version Bolt could run with to 2.x,
as we didn't know if there would be breaking changes in Ruby 3. This commit updates
the gemspec to no longer confine to ruby 2.  Bolt will continue to ship with Ruby 2.7, so
this only applies to gem installations. This adds test matrices for Ruby
3 to the unit and integration tests.

!feature

* **Bolt can now be run with Ruby 3** ([#2767](https://github.com/puppetlabs/bolt/issues/2767))
  Required Ruby version for running Bolt is now simply greater than 2.5,
  rather than between 2.5 and 3.0, allowing Bolt gem installs to run with
  Ruby 3.