(GH-2767) Allow Bolt to run with Ruby 3.0
Previously, our gemspec limited the Ruby version Bolt could run with to 2.x,
as we didn't know if there would be breaking changes in Ruby 3. Now
that 3 has been released and there don't appear to be any changes that
will make Bolt unusable with Ruby 3, we've modified the gemspec to allow
any Ruby greater than 2.5. Bolt will continue to ship with Ruby 2.7, so
this only applies to gem installations. This adds test matrices for Ruby
3 to the unit and integration tests.

!feature

* **Bolt can now be run with Ruby 3** ([#2767](https://github.com/puppetlabs/bolt/issues/2767))
  Required Ruby version for running Bolt is now simply greater than 2.5,
  rather than between 2.5 and 3.0, allowing Bolt gem installs to run with
  Ruby 3.