(GH-2771) Add puppetdb_command plan function
This adds a new `puppetdb_command` plan function which can be used to
invoke commands in PuppetDB. The function accepts three parameters: the
command, the command version, and the payload. Bolt uses these values to
build a POST request to the command v1 endpoint for PuppetDB. If the
request is successful, the command returns the UUID for the command
invocation, which can be used to reference the command in logs.

Currently, this function only supports and tests the `replace_facts`
command, though other commands might still work.

!feature

* **`puppetdb_command` plan function**
  ([#2771](https://github.com/puppetlabs/bolt/issues/2771))

  The `puppetdb_command` plan function can be used to invoke commands in
  PuppetDB. Currently, only the `replace_facts` command is officially
  tested and supported, though other commands might work as well.

  _This feature is experimental and subject to change._