(GH-2764) Add 'background' and 'wait' plan functions
This fundamentally changes how plans are run in Bolt in order to support
two new plan functions, `background()` and `wait()`. Previously, users
who wanted to parallelize a Bolt plan had to use the `parallelize()`
plan function, which was built for fan-out parallelization. With
`parallelize()`, fire-and-forget parallelization was difficult because
users would have to include "the rest of the plan" in the body of the
`parallelize()` block, which was hard to read and understand.

To support fire-and-forget parallelism this change introduces
`background()` and `wait()` plan functions. `background()` runs the
provided block in parallel with the rest of the plan, and does not block
unless `wait()` is called. `wait()` blocks on the results of a
backgrounded block and returns them. `background()` creates a new
Bolt::Future object, which `wait()` accepts as an argument.

With `parallelize()` the `wait()` functionality was builtin to the
function, so the main plan itself could be executed serially with only
the parallelized blocks running in parallel (along with the main plan).
However with the `background()` function the main plan needs to run in
parallel with backgrounded blocks.

Bolt now starts the main plan in a Fiber and continuously resumes that
Fiber, even when there are no other blocks running (which could be a
future improvement). This shouldn't effect serial plan execution, and
means that Bolt doesn't need to have a serial and parallel execution
mode - it just always executes in "parallel" even if there's only one
"thread".

!feature

* **Add 'background()' and 'wait()' plan functions** ([#2764](https://github.com/puppetlabs/bolt/issues/2764))
  This introduces two new plan functions, `background()` and `wait()`,
  to support fire-and-forget parallelism in Bolt. Read more at
  https://pup.pt/bolt-parallelism.